Move container code to its own file
