+++ src/ui/index.js
@@ -1,1 +1,2 @@
+export {default as addContainerDiv} from "./container";
 export {default as injectStyling} from "./styling";
